+++ app/src/main/java/woowacourse/paint/customview/CanvasView.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint
+package woowacourse.paint.customview
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -6,6 +6,10 @@ import android.graphics.Canvas
 import android.util.AttributeSet
 import android.view.MotionEvent
 import android.view.View
+import woowacourse.paint.BrushHistory
+import woowacourse.paint.Color
+import woowacourse.paint.brush.Brush
+import woowacourse.paint.brush.LineBrush
 
 class CanvasView constructor(context: Context, attr: AttributeSet) : View(context, attr) {
