+++ app/src/main/res/layout/activity_main.xml
@@ -15,7 +15,7 @@
         app:layout_constraintTop_toTopOf="parent"
         app:titleTextColor="?attr/colorOnPrimary" />
 
-    <woowacourse.paint.CanvasView
+    <woowacourse.paint.customview.CanvasView
         android:id="@+id/canvasView"
         android:layout_width="0dp"
         android:layout_height="0dp"
@@ -69,7 +69,7 @@
         android:valueTo="100"
         app:layout_constraintBottom_toTopOf="@id/paletteView" />
 
-    <woowacourse.paint.PaletteView
+    <woowacourse.paint.customview.PaletteView
         android:id="@+id/paletteView"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
